docker network fun

+++ Makefile
@@ -1,9 +1,9 @@
 start:
-	docker network create chuntfm-liquidsoap
+	if ! docker network ls | grep -q chuntfm-liquidsoap; then docker network create chuntfm-liquidsoap; fi
 	docker-compose up -d
 
 stop:
 	docker-compose down
-	docker network rm chuntfm-liquidsoap
+	if docker network ls | grep -q chuntfm-liquidsoap; then docker network rm chuntfm-liquidsoap; fi
 clean:
 	docker rmi nginx:alpine
